+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Add `trim` method
 - Add `replaceBetween` method
+- Add `removeNewLine` method
 
 ## 0.0.1 (2024-09-02) - Not for Production